feat(web): remove org scope from host inventory
