+++ apps/web/app/(dashboard)/hosts/networks/components/host-node-terminal-dialog.tsx
@@ -52,7 +52,6 @@ function TerminalConnectForm({
       hostname: data.name,
       username: username.trim(),
       password,
-      orgId: data.orgId,
       directAccess: false,
     })
     setPassword('')
