+++ apps/web/components/shared/command-palette/command-palette.tsx
@@ -64,12 +64,11 @@ function writeRecents(ids: string[]): void {
 }
 
 interface CommandPaletteProviderProps {
-  orgId: string
   userRole: string
   children: React.ReactNode
 }
 
-export function CommandPaletteProvider({ orgId, userRole, children }: CommandPaletteProviderProps) {
+export function CommandPaletteProvider({ userRole, children }: CommandPaletteProviderProps) {
   const [isOpen, setIsOpen] = useState(false)
 
   const open = useCallback(() => setIsOpen(true), [])
@@ -97,22 +96,21 @@ export function CommandPaletteProvider({ orgId, userRole, children }: CommandPal
   return (
     <CommandPaletteContext.Provider value={value}>
       {children}
-      <CommandPaletteDialog orgId={orgId} userRole={userRole} isOpen={isOpen} onOpenChange={setIsOpen} />
+      <CommandPaletteDialog userRole={userRole} isOpen={isOpen} onOpenChange={setIsOpen} />
     </CommandPaletteContext.Provider>
   )
 }
 
 interface CommandPaletteDialogProps {
-  orgId: string
   userRole: string
   isOpen: boolean
   onOpenChange: (next: boolean) => void
 }
 
-function CommandPaletteDialog({ orgId, userRole, isOpen, onOpenChange }: CommandPaletteDialogProps) {
+function CommandPaletteDialog({ userRole, isOpen, onOpenChange }: CommandPaletteDialogProps) {
   const router = useRouter()
   const navItems = useNavigationItems(userRole)
-  const hostItems = useHostItems(orgId, isOpen)
+  const hostItems = useHostItems(isOpen)
 
   const allItems = useMemo(() => [...navItems, ...hostItems], [navItems, hostItems])
