+++ apps/web/hooks/use-host-stream.ts
@@ -6,7 +6,7 @@ import type { HostWithAgent } from '@/lib/actions/agents'
 import type { CheckWithHistory } from '@/lib/actions/checks'
 import type { ResolvedNote } from '@/lib/actions/notes-resolver'
 
-export function useHostStream({ hostId, orgId }: { hostId: string; orgId: string }) {
+export function useHostStream({ hostId }: { hostId: string; [key: string]: unknown }) {
   const queryClient = useQueryClient()
   const esRef = useRef<EventSource | null>(null)
 
@@ -19,7 +19,7 @@ export function useHostStream({ hostId, orgId }: { hostId: string; orgId: string
     es.addEventListener('update', (e) => {
       try {
         const host: HostWithAgent = JSON.parse(e.data)
-        queryClient.setQueryData(['host', orgId, hostId], host)
+        queryClient.setQueryData(['host', hostId], host)
       } catch {
         // malformed JSON — ignore
       }
@@ -28,7 +28,7 @@ export function useHostStream({ hostId, orgId }: { hostId: string; orgId: string
     es.addEventListener('checks', (e) => {
       try {
         const checks: CheckWithHistory[] = JSON.parse(e.data)
-        queryClient.setQueryData(['checks-history', orgId, hostId], checks)
+        queryClient.setQueryData(['checks-history', hostId], checks)
       } catch {
         // malformed JSON — ignore
       }
@@ -49,7 +49,7 @@ export function useHostStream({ hostId, orgId }: { hostId: string; orgId: string
           updatedAt: new Date(n.updatedAt),
           deletedAt: n.deletedAt ? new Date(n.deletedAt) : null,
         }))
-        queryClient.setQueryData(['notes-for-host', orgId, hostId], notes)
+        queryClient.setQueryData(['notes-for-host', hostId], notes)
       } catch {
         // malformed JSON — ignore
       }
@@ -59,5 +59,5 @@ export function useHostStream({ hostId, orgId }: { hostId: string; orgId: string
       esRef.current?.close()
       esRef.current = null
     }
-  }, [hostId, orgId, queryClient])
+  }, [hostId, queryClient])
 }
