+++ apps/web/lib/actions/agents-delete-hard.test.mjs
@@ -5,11 +5,11 @@ import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 
 const here = path.dirname(fileURLToPath(import.meta.url))
-const agentsSource = readFileSync(path.join(here, 'agents.ts'), 'utf8')
+const agentsSource = readFileSync(path.join(here, 'agents-core.ts'), 'utf8')
 
 function actionSegment(action) {
-  const start = agentsSource.indexOf(`export async function ${action}`)
-  assert.notEqual(start, -1, `expected ${action} to exist in agents.ts`)
+  const start = agentsSource.lastIndexOf(`export async function ${action}`)
+  assert.notEqual(start, -1, `expected ${action} to exist in agents-core.ts`)
   const next = agentsSource.indexOf('\nexport async function ', start + 1)
   return agentsSource.slice(start, next === -1 ? undefined : next)
 }
