+++ apps/web/lib/actions/lifecycle-audit.test.mjs
@@ -5,11 +5,11 @@ import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 
 const here = path.dirname(fileURLToPath(import.meta.url))
-const agentsSource = readFileSync(path.join(here, 'agents.ts'), 'utf8')
+const agentsSource = readFileSync(path.join(here, 'agents-core.ts'), 'utf8')
 const usersSource = readFileSync(path.join(here, 'users.ts'), 'utf8')
 
 function getActionSegment(source, action) {
-  const start = source.indexOf(`export async function ${action}`)
+  const start = source.lastIndexOf(`export async function ${action}`)
   assert.notEqual(start, -1, `expected ${action} to exist`)
   const next = source.indexOf('\nexport async function ', start + 1)
   return source.slice(start, next === -1 ? undefined : next)
@@ -20,9 +20,9 @@ const auditedLifecycleActions = [
   ['users.ts', usersSource, 'deactivateUser', ['user.deactivated']],
   ['users.ts', usersSource, 'reactivateUser', ['user.reactivated']],
   ['users.ts', usersSource, 'cancelInvite', ['invitation.cancelled']],
-  ['agents.ts', agentsSource, 'approveAgent', ['agent.approved']],
-  ['agents.ts', agentsSource, 'rejectAgent', ['agent.rejected']],
-  ['agents.ts', agentsSource, 'revokeEnrolmentToken', ['agent.enrolment_token.revoked']],
+  ['agents-core.ts', agentsSource, 'approveAgent', ['agent.approved']],
+  ['agents-core.ts', agentsSource, 'rejectAgent', ['agent.rejected']],
+  ['agents-core.ts', agentsSource, 'revokeEnrolmentToken', ['agent.enrolment_token.revoked']],
 ]
 
 test('privileged lifecycle mutations write central audit events', () => {
